Update getting started for AMD users.
As I faced problem in setting up the react-native environment it is hard to understand where the problem is. Most of the time we tell on the basis of Operating system but sometimes all the things are not working same in all the processors. Here 'android-studio' is not working correctly. The android emulator is not able to open as there is no intel HAXM. This is simple update for the AMD users.