Add section : Known Issues with fetch and [..] (#769)
* Add section : Known Issues with fetch and [..]

Cookie based authentication is currently wobbly. Had it not been for the problematic redirect scenario using fetch workarounds could be found (e.g clearing cookies before every single request and manually adding them to the headers). However, because of the 302 issue there is a scenario where using cookies can result in halt of development.

The rest of the issues are merely added here so that people that can follow workarounds ( please take a look as unintended consequences might occur )

* Minor text changes and remove the last paragraph.